# runtime/swagger/http.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MineAdmin API DOC",
        "description": "MineAdmin后台接口列表",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9501",
            "description": "本地服务器"
        }
    ],
    "paths": {
...
        "/system/app/index": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/recycle": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/save": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/update/{id}": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/delete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/realDelete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/recovery": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/changeStatus": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/remote": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/role/index": {
            "get": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/recycle": {
            "get": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/getMenuByRole/{id}": {
            "get": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleGetMenuByRole{id}GET",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/getDeptByRole/{id}": {
            "get": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleGetDeptByRole{id}GET",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/list": {
            "get": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleListGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/save": {
            "post": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/update/{id}": {
            "put": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/menuPermission/{id}": {
            "put": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleMenuPermission{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/dataPermission/{id}": {
            "put": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleDataPermission{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/delete": {
            "delete": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/realDelete": {
            "delete": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/recovery": {
            "put": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/changeStatus": {
            "put": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/numberOperation": {
            "put": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleNumberOperationPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/role/remote": {
            "post": {
                "tags": [
                    "RoleController"
                ],
                "summary": "",
                "operationId": "SystemRoleRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}